requirements.txt
fastapi
uvicorn
Pillow
numpy
rembg
python-multipart